Add clang-tidy bugprone checks (#751)
* bugprone-narrowing-conversions

* Add missing reference

* bugprone-unchecked-optional-access

* bugprone-empty-catch

* Start on next check.

* Build fix

* Fix some more easily swappable parameters.

* More build fixes

* Even more build fixes.

* PosssiblePath swappables

* Fix build

* Clear up ship construction

* Strategic regions

* OutWorld

* Partial railways_converter

* Build fix

* More railways.

* Build fix

* Turn off linting for parts of localizations_converter

* Remove some unneeded functions.

* Revert options change

* Formatting

* Fix test

* Build fix

* Try turning on one last check

* Yeah, that's giving hundreds of false positives, no thanks.

* Warning as error

+++ src/hoi4_world/military/convoy_distributor.cpp
@@ -48,7 +48,7 @@ int ConvoyDistributor::ConvoysFromState(int id) const
       return 0;
    }
    auto weight = vic3_map_.at(id);
-   weight *= hoi4_total_;
+   weight *= static_cast<float>(hoi4_total_);
    weight /= vic3_total_;
    return static_cast<int>(std::round(weight));
 }
